feat: implement all the tests: track, album, artist without errors. Implement 53 tests

+++ src/albums/albums.controller.ts
@@ -1,4 +1,5 @@
 import { Controller, Get, Post, Body, Param, Delete, HttpCode, Put, Res, HttpException } from '@nestjs/common';
+import { isValidationOptions } from 'class-validator';
 import { Response } from 'express';
 import { StatusCodes } from 'http-status-codes';
 import { validate } from 'uuid';
@@ -32,9 +33,9 @@ export class AlbumsController {
   @Get(':id')
   @HttpCode(200)
   findOne(@Param('id') id: string, @Res() res: Response) {
-    if (typeof id !== "string") {
-      throw new HttpException(`id ${id} not validate`, StatusCodes.BAD_REQUEST);
-    }
+    // if (typeof id !== "string") {
+    //   throw new HttpException(`id ${id} not validate`, StatusCodes.BAD_REQUEST);
+    // }
     if (!validate(id)) {
       throw new HttpException(`id ${id} not validate UUID`, StatusCodes.BAD_REQUEST);
     }
@@ -48,14 +49,36 @@ export class AlbumsController {
 
   @Put(':id')
   @HttpCode(200)
-  update(@Param('id') id: string, @Body() updateAlbumDto: UpdateAlbumDto) {
-    return this.albumsService.update(id, updateAlbumDto);
+  update(@Param('id') id: string, @Body() updateAlbumDto: UpdateAlbumDto, @Res() res: Response) {
+    if (typeof id !== "string") {
+      throw new HttpException(`id ${id} not validate`, StatusCodes.BAD_REQUEST);
+    }
+    if (!validate(id)) {
+      throw new HttpException(`id ${id} not validate UUID`, StatusCodes.BAD_REQUEST);
+    }
+    // if (!updateAlbumDto) {
+    //   throw new HttpException(`Not validate type of data`, StatusCodes.BAD_REQUEST);
+    // }
+    const album = this.albumsService.update(id, updateAlbumDto);
+    if (!album) {
+      throw new HttpException(`Album not found`, StatusCodes.NOT_FOUND);
+    }
+    res.send(album);
+    return album;
   }
 
   @Delete(':id')
   @HttpCode(204)
-  remove(@Param('id') id: string) {
-    return this.albumsService.remove(id);
+  remove(@Param('id') id: string, @Res() res: Response) {
+    if (!validate(id)) {
+      throw new HttpException(`id ${id} not validate UUID`, StatusCodes.BAD_REQUEST);
+    }
+    let album = this.albumsService.remove(id);
+    if (!album) {
+      throw new HttpException(`Album not found`, StatusCodes.NOT_FOUND);
+    }
+    res.send(StatusCodes.NO_CONTENT);
+    return;
   }
 }
